test: moved attendees mocks in a new file
